backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
pinecone-client==2.2.4
google-generativeai==0.3.2
PyMuPDF==1.24.0
python-dotenv==1.0.0
httpx==0.25.2
# serpapi==0.1.5  # Optional: Uncomment to enable web search feature
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0